create custom view for morecard
